pyproject.toml
[project]
name = "browseruse-cdp-best-practise"
version = "0.1.0"
description = "CDP + Playwright + BrowserUse Integration Best Practise"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "browser-use>=0.11.0",
    "playwright>=1.48.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
